requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.14
django-allauth==0.40.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-pyodbc-azure==2.1.0.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
pyodbc==4.0.27
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
